requirement.txt
phidata
groq
yfinance
'fastapi[standard]'
sqlalchemy
duckduckgo-search